+++ jaxley/synapses/__init__.py
@@ -3,7 +3,14 @@
 
 from jaxley.synapses.ionotropic import IonotropicSynapse
 from jaxley.synapses.synapse import Synapse
+from jaxley.synapses.tanh_conductance import TanhConductanceSynapse
 from jaxley.synapses.tanh_rate import TanhRateSynapse
 from jaxley.synapses.test import TestSynapse
 
-__all__ = ["IonotropicSynapse", "Synapse", "TanhRateSynapse", "TestSynapse"]
+__all__ = [
+    "IonotropicSynapse",
+    "Synapse",
+    "TanhConductanceSynapse",
+    "TanhRateSynapse",
+    "TestSynapse",
+]